add web and io implemenations of network and asset image
