+++ packages/flutter/test/painting/image_resolution_test.dart
@@ -10,7 +10,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/painting.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-
+import 'package:flutter/src/painting/_image_resolution_io.dart' as io;
 
 class TestAssetBundle extends CachingAssetBundle {
   TestAssetBundle(this._assetBundleMap);
@@ -44,7 +44,7 @@ void main() {
 
       assetBundleMap[mainAssetPath] = <String>[];
 
-      final AssetImage assetImage = AssetImage(
+      final io.AssetImage assetImage = io.AssetImage(
           mainAssetPath,
           bundle: TestAssetBundle(assetBundleMap));
       const ImageConfiguration configuration = ImageConfiguration();
@@ -95,7 +95,7 @@ void main() {
       final TestAssetBundle testAssetBundle = TestAssetBundle(
           assetBundleMap);
 
-      final AssetImage assetImage = AssetImage(
+      final io.AssetImage assetImage = io.AssetImage(
           mainAssetPath,
           bundle: testAssetBundle);
 
@@ -128,7 +128,7 @@ void main() {
       final TestAssetBundle testAssetBundle = TestAssetBundle(
           assetBundleMap);
 
-      final AssetImage assetImage = AssetImage(
+      final io.AssetImage assetImage = io.AssetImage(
           mainAssetPath,
           bundle: TestAssetBundle(assetBundleMap));
 
@@ -167,7 +167,7 @@ void main() {
       final TestAssetBundle testAssetBundle = TestAssetBundle(
           assetBundleMap);
 
-      final AssetImage assetImage = AssetImage(
+      final io.AssetImage assetImage = io.AssetImage(
           mainAssetPath,
           bundle: testAssetBundle);
